Extract cache dependency classes to separate files.
